+++ game_logic.py
@@ -1,69 +1,3 @@
-import random
-
-class Player:
-    points = 0
-    dominoes = []
-
-    def __init__(self, id):
-        self.id = id
-
-    #setting how the player_class object will be displayed
-    def __str__(self):
-        return "Player {}".format(self.id + 1)
-
-class Domino:
-
-    def __init__(self, top_value, bottom_value):
-        self.top_value = top_value
-        self.bottom_value = bottom_value
-
-    #setting how the domino_class object will be displayed
-    def __str__(self):
-        return "[{}|{}]".format(self.top_value, self.bottom_value)
-
-class Game_data:
-
-    def __init__(self, player_class, domino_class):
-        self.player_class = player_class
-        self.domino_class = domino_class
-
-    def create_dominoes(self):
-        dominoes = []
-        for i in range(7):
-            for j in range(i,7):        
-                dominoes.append(self.domino_class(i, j))
-        return dominoes
-
-    @staticmethod
-    def shuffle_dominoes(dominoes):
-        for i in range(len(dominoes)):
-            random_position = random.randrange(len(dominoes))
-            dominoes[i], dominoes[random_position] = dominoes[random_position], dominoes[i]
-
-    def create_players(self, number_of_players):
-        players = []
-        for i in range(number_of_players):
-            players.append(self.player_class(i))
-        return players
-
-    @staticmethod
-    def distribute_dominoes(dominoes, players):
-        for player in players:
-            player.dominoes = dominoes[0:7]
-            dominoes = dominoes[7:]
-        return dominoes
-
-    def get_game_data(self, number_of_players):
-        #creating dominoes
-        dominoes = self.create_dominoes()
-        #shuffling dominoes
-        self.shuffle_dominoes(dominoes)
-        #creating players objects
-        players = self.create_players(number_of_players)
-        #distribute dominoes
-        dominoes = self.distribute_dominoes(dominoes, players)
-        return players, dominoes
-
 class Game_logic:
     turn = 0
     game_table = []
@@ -282,35 +216,3 @@ class Game_logic:
         self.players, self.dominoes = self.game_data.get_game_data(self.number_of_players)
         self.first_move()
         self.game_loop()
-
-
-def number_of_players_validation():
-  while(True):
-    number_of_players = input()
-    if len(number_of_players) == 1 and number_of_players.isdigit() and (int(number_of_players) > 1 and int(number_of_players) < 5):
-      return int(number_of_players)
-    else:
-      print("wrong input")
-
-def points_to_win_validation():
-  while(True):
-    points_to_win = input()
-    if len(points_to_win) > 0 and points_to_win.isdigit() and int(points_to_win) > 0:
-      return int(points_to_win)
-    else:
-      print("Wrong input")
-
-def game_setup():
-    print('how many players are going to play?')
-    number_of_players = number_of_players_validation()
-    print("what's limit of points to win that you want to set?")
-    points_to_win = points_to_win_validation()
-    #create game data object
-    game_data = Game_data(Player, Domino)
-    #create game logic object 
-    game_logic = Game_logic(game_data, points_to_win, number_of_players)
-    #start the game
-    game_logic.game_start()
-
-game_setup()
-
